DevOps-project/yolo/requirements.txt
flask
pyyaml
loguru
boto3
pymongo
# testing
pymongo
Flask-PyMongo
pylint
pytest
